Reorganize theme2 crate
