+++ crates/theme2/src/styles/colors.rs
@@ -2,54 +2,7 @@ use gpui::Hsla;
 use refineable::Refineable;
 use std::sync::Arc;
 
-use crate::{PlayerColors, SyntaxTheme, SystemColors};
-
-#[derive(Refineable, Clone, Debug)]
-#[refineable(Debug, serde::Deserialize)]
-pub struct StatusColors {
-    /// Indicates some kind of conflict, like a file changed on disk while it was open, or
-    /// merge conflicts in a Git repository.
-    pub conflict: Hsla,
-
-    /// Indicates something new, like a new file added to a Git repository.
-    pub created: Hsla,
-
-    /// Indicates that something no longer exists, like a deleted file.
-    pub deleted: Hsla,
-
-    /// Indicates a system error, a failed operation or a diagnostic error.
-    pub error: Hsla,
-
-    /// Represents a hidden status, such as a file being hidden in a file tree.
-    pub hidden: Hsla,
-
-    /// Indicates a hint or some kind of additional information.
-    pub hint: Hsla,
-
-    /// Indicates that something is deliberately ignored, such as a file or operation ignored by Git.
-    pub ignored: Hsla,
-
-    /// Represents informational status updates or messages.
-    pub info: Hsla,
-
-    /// Indicates a changed or altered status, like a file that has been edited.
-    pub modified: Hsla,
-
-    /// Indicates something that is predicted, like automatic code completion, or generated code.
-    pub predictive: Hsla,
-
-    /// Represents a renamed status, such as a file that has been renamed.
-    pub renamed: Hsla,
-
-    /// Indicates a successful operation or task completion.
-    pub success: Hsla,
-
-    /// Indicates some kind of unreachable status, like a block of code that can never be reached.
-    pub unreachable: Hsla,
-
-    /// Represents a warning status, like an operation that is about to fail.
-    pub warning: Hsla,
-}
+use crate::{PlayerColors, StatusColors, SyntaxTheme, SystemColors};
 
 #[derive(Refineable, Clone, Debug)]
 #[refineable(Debug, serde::Deserialize)]
@@ -283,7 +236,7 @@ mod tests {
 
     #[test]
     fn override_a_single_theme_color() {
-        let mut colors = ThemeColors::default_light();
+        let mut colors = ThemeColors::light();
 
         let magenta: Hsla = gpui::rgb(0xff00ff);
 
@@ -301,7 +254,7 @@ mod tests {
 
     #[test]
     fn override_multiple_theme_colors() {
-        let mut colors = ThemeColors::default_light();
+        let mut colors = ThemeColors::light();
 
         let magenta: Hsla = gpui::rgb(0xff00ff);
         let green: Hsla = gpui::rgb(0x00ff00);
