+++ crates/theme2/src/styles/system.rs
@@ -1,4 +1,4 @@
-use gpui::Hsla;
+use gpui::{hsla, Hsla};
 
 #[derive(Clone)]
 pub struct SystemColors {
@@ -6,4 +6,15 @@ pub struct SystemColors {
     pub mac_os_traffic_light_red: Hsla,
     pub mac_os_traffic_light_yellow: Hsla,
     pub mac_os_traffic_light_green: Hsla,
+}
+
+impl Default for SystemColors {
+    fn default() -> Self {
+        Self {
+            transparent: hsla(0.0, 0.0, 0.0, 0.0),
+            mac_os_traffic_light_red: hsla(0.0139, 0.79, 0.65, 1.0),
+            mac_os_traffic_light_yellow: hsla(0.114, 0.88, 0.63, 1.0),
+            mac_os_traffic_light_green: hsla(0.313, 0.49, 0.55, 1.0),
+        }
+    }
 }
